+++ xcube_cds/datasets/drought_indices_era5.py
@@ -24,6 +24,7 @@ import glob
 import os
 import pathlib
 
+import numpy as np
 import pandas as pd
 import xarray as xr
 from xcube.core.store import DatasetDescriptor, VariableDescriptor
@@ -44,6 +45,10 @@ class DroughtIndicesDatasetHandler(CDSDatasetHandler):
             "derived-drought-historical-monthly:reanalysis": (
                 "Monthly drought indices from 1940–present derived "
                 "from ERA5 reanalysis (main run)"
+            ),
+            "derived-drought-historical-monthly:ensemble_members": (
+                "Monthly drought indices from 1940–present derived "
+                "from ERA5 ensemble (10 members)"
             ),
         }
         self._variable_names = [
@@ -114,13 +119,18 @@ class DroughtIndicesDatasetHandler(CDSDatasetHandler):
                     var_name, accum_period
                 )
 
+        if data_id.endswith("reanalysis"):
+            dims = ("time", "lat", "lon")
+        else:
+            dims = ("time", "number", "lat", "lon")
+
         variable_descriptors = []
         for var_name, attrs in mapping_varname_attrs.items():
             variable_descriptors.append(
                 VariableDescriptor(
                     name=var_name,
                     dtype="float64",
-                    dims=("time", "lat", "lon"),
+                    dims=dims,
                     attrs=attrs,
                 )
             )
@@ -177,33 +187,72 @@ class DroughtIndicesDatasetHandler(CDSDatasetHandler):
             zip_ref.extractall(path_temp)
         file_paths = glob.glob(f"{path_temp}/*")
         dss = []
-        for var_name in open_params["variable_names"]:
-            for accum_period in open_params["accumulation_periods"]:
-                pattern = self._get_filepath_pattern(var_name, accum_period)
-                file_sel = [path for path in file_paths if pattern in path]
-                file_sel = sorted(file_sel)
-                ds = xr.open_mfdataset(
-                    file_sel,
-                    engine="netcdf4",
-                    chunks="auto",
-                    combine_attrs="drop_conflicts",
-                )
-                if "standardised_precipitation" in var_name:
-                    ds = ds.sel(
-                        time=slice(
-                            open_params["time_range"][0], open_params["time_range"][1]
-                        )
+
+        if cds_api_params["product_type"] == ["reanalysis"]:
+            for var_name in open_params["variable_names"]:
+                for accum_period in open_params["accumulation_periods"]:
+                    pattern = self._get_filepath_pattern(var_name, accum_period)
+                    file_sel = [path for path in file_paths if pattern in path]
+                    file_sel = sorted(file_sel)
+                    ds = xr.open_mfdataset(
+                        file_sel,
+                        engine="netcdf4",
+                        chunks="auto",
+                        combine_attrs="drop_conflicts",
                     )
-                else:
-                    ds = self._resample_quality_ds(ds, open_params["time_range"])
-                assert len(ds.data_vars) == 1
-                ds_varname = self._get_varname(var_name, accum_period)
-                ds = ds.rename({list(ds.data_vars.keys())[0]: ds_varname})
-                dss.append(ds)
-        ds_final = xr.merge(dss, join="outer", combine_attrs="drop_conflicts")
-        ds_final = ds_final.sel(
-            time=slice(open_params["time_range"][0], open_params["time_range"][1])
-        )
+                    if "standardised_precipitation" in var_name:
+                        ds = ds.sel(
+                            time=slice(
+                                open_params["time_range"][0],
+                                open_params["time_range"][1],
+                            )
+                        )
+                    else:
+                        ds = self._resample_quality_ds(ds, open_params["time_range"])
+                    assert len(ds.data_vars) == 1
+                    ds_varname = self._get_varname(var_name, accum_period)
+                    ds = ds.rename({list(ds.data_vars.keys())[0]: ds_varname})
+                    dss.append(ds)
+            ds_final = xr.merge(dss, join="outer", combine_attrs="drop_conflicts")
+            ds_final = ds_final.sel(
+                time=slice(open_params["time_range"][0], open_params["time_range"][1])
+            )
+        else:
+            for var_name in open_params["variable_names"]:
+                for accum_period in open_params["accumulation_periods"]:
+                    pattern = self._get_filepath_pattern(var_name, accum_period)
+                    file_sel = [path for path in file_paths if pattern in path]
+                    file_sel = sorted(file_sel)
+                    dss_inner = []
+                    for path in file_sel:
+                        ds = xr.open_dataset(path, engine="netcdf4", chunks="auto")
+                        time_axis = ds.time
+                        # The data from the backend uses the confusing name `time` for the
+                        # ensemble member index. We rename it to `number` to be consistent
+                        # with other ERA5 datasets, and to free up the name `time` for the actual
+                        # time.
+                        ds = ds.rename({"time": "number"})
+                        ds = ds.assign_coords(number=np.arange(10))
+                        ds = ds.expand_dims(time=[time_axis[0].values])
+                        dss_inner.append(ds)
+                    ds = xr.concat(dss_inner, "time", combine_attrs="drop_conflicts")
+                    if "standardised_precipitation" in var_name:
+                        ds = ds.sel(
+                            time=slice(
+                                open_params["time_range"][0],
+                                open_params["time_range"][1],
+                            )
+                        )
+                    else:
+                        ds = self._resample_quality_ds(ds, open_params["time_range"])
+                    assert len(ds.data_vars) == 1
+                    ds_varname = self._get_varname(var_name, accum_period)
+                    ds = ds.rename({list(ds.data_vars.keys())[0]: ds_varname})
+                    dss.append(ds)
+            ds_final = xr.merge(dss, join="outer", combine_attrs="drop_conflicts")
+            ds_final = ds_final.sel(
+                time=slice(open_params["time_range"][0], open_params["time_range"][1])
+            )
         return ds_final
 
     @staticmethod