+++ CHANGES.md
@@ -2,6 +2,7 @@
 
  - Adjust time range in ERA5 demo notebook. (#106)
  - Add data access for [monthly drought indices from 1940 to present derived from ERA5 reanalysis](https://cds.climate.copernicus.eu/datasets/derived-drought-historical-monthly?tab=overview). (#109)
+ - Add data access for [monthly drought indices from 1940 to present derived from ERA5 ensemble members](https://cds.climate.copernicus.eu/datasets/derived-drought-historical-monthly?tab=overview). (#110)
  - Support chunking of returned datasets. (#110)
 
 ## Changes in 1.1.0